backend/pyproject.toml
[tool.poetry]
name = "portfolio-backend"
version = "0.1.0"
description = "Django backend for Portfolio project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
dj-database-url = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
